requirements.txt
streamlit==1.35.0
python-dotenv==1.0.1
kafka-python==2.0.2
google-api-python-client==2.121.0
textblob==0.17.1
transformers==4.40.1
torch==2.3.0
azure-ai-textanalytics==5.3.0
azure-core==1.30.1
matplotlib==3.8.4
wordcloud==1.9.3
nltk==3.8.1
